+++ src/molecules/checkbox-input/checkbox-input.scss
