+++ src/molecules/form-fields/select-form-field.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import uuid from "uuid";
-import { Select, SelectOption } from "../../atoms/forms/select";
+import { Select, SelectOption } from "../select/select";
 import { StringUtils, CollectionUtils } from "andculturecode-javascript-core";
 import "./form-field.scss";
