+++ src/molecules/progress-bar/progress-bar.scss
