+++ src/molecules/radio-input/radio-input.scss
@@ -1,5 +1,5 @@
 @import "../../assets/scss/6-components/_component-base";
-@import "../buttons/button.scss";
+@import "../../Atoms/buttons/button.scss"; // Todo: Remove this dependency
 
 .c-radio {
     @include padding(16px);
